refactor: move validateOptions to settting
