+++ pkg/juicefs/juicefs_test.go
@@ -1213,106 +1213,6 @@ func Test_juicefs_validTarget(t *testing.T) {
 	}
 }
 
-func Test_juicefs_validOptions(t *testing.T) {
-	type args struct {
-		volumeId string
-		options  []string
-		volCtx   map[string]string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    []string
-		wantErr bool
-	}{
-		{
-			name: "test-normal",
-			args: args{
-				volumeId: "test",
-				options:  []string{"cache-dir=xxx"},
-			},
-			want:    []string{"cache-dir=xxx"},
-			wantErr: false,
-		},
-		{
-			name: "test-space1",
-			args: args{
-				volumeId: "test",
-				options:  []string{" cache-dir=xxx "},
-			},
-			want:    []string{"cache-dir=xxx"},
-			wantErr: false,
-		},
-		{
-			name: "test-space2",
-			args: args{
-				volumeId: "test",
-				options:  []string{" cache-dir = xxx "},
-			},
-			want:    []string{"cache-dir=xxx"},
-			wantErr: false,
-		},
-		{
-			name: "test-error",
-			args: args{
-				volumeId: "test",
-				options:  []string{"cache-dir=xxx cache-size=1024"},
-			},
-			want:    []string{},
-			wantErr: true,
-		},
-		{
-			name: "test-buffersize",
-			args: args{
-				volumeId: "test",
-				options:  []string{"buffer-size=1024"},
-				volCtx: map[string]string{
-					config.MountPodMemLimitKey: "1Mi",
-				},
-			},
-			want:    []string{},
-			wantErr: true,
-		},
-		{
-			name: "test-buffersize-with-unit",
-			args: args{
-				volumeId: "test",
-				options:  []string{"buffer-size=1024M"},
-				volCtx: map[string]string{
-					config.MountPodMemLimitKey: "1Mi",
-				},
-			},
-			want:    []string{},
-			wantErr: true,
-		},
-		{
-			name: "test-buffersize-with-unit",
-			args: args{
-				volumeId: "test",
-				options:  []string{"buffer-size=10M"},
-				volCtx: map[string]string{
-					config.MountPodMemLimitKey: "20Mi",
-				},
-			},
-			want:    []string{"buffer-size=10M"},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			j := &juicefs{}
-			got, err := j.validOptions(tt.args.volumeId, tt.args.options, tt.args.volCtx)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("validOptions() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("validOptions() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func Test_parseRawVersion(t *testing.T) {
 	type parseRawVersionCase struct {
 		name string
